+++ index.html
@@ -74,7 +74,7 @@
   <script src="js/utils/list.min.js"></script>
 
   <script>
-    var app = require('./js/app'),
+    var app = require('./js/main'),
         remote = require('remote');
     $('.js-closeIndexModal').on('click', function(){
       $(this).closest('.modal').addClass('hide');
