+++ js/views/serverConnectModal.js
@@ -5,7 +5,7 @@ var __ = require('underscore'),
     loadTemplate = require('../utils/loadTemplate'),
     isLocalServerRunning = require('../utils/isLocalServerRunning'),    
     isRemoteServerRunning = require('../utils/isRemoteServerRunning'),
-    app = require('../App2.js').getApp(),        
+    app = require('../App.js').getApp(),        
     serverConfigMd = require('../models/serverConfigMd'),
     baseModal = require('./baseModal'),
     messageModal = require('../utils/messageModal.js');
